requirements.txt
fastapi
uvicorn[standard]
requests
packaging
aiohttp
httpx
Pillow
psutil
aiosqlite
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
user-agents
